type-c-service: Break up wrapper struct into submodules
Break up this struct into submodules to help keep things readable.
