+++ type-c-service/src/wrapper/mod.rs
@@ -5,12 +5,14 @@ use core::cell::{Cell, RefCell};
 
 use bitfield::BitMut;
 use embassy_futures::select::{select, select_array, Either};
-use embedded_services::power::policy::device::{RequestData, StateKind};
+use embedded_services::power::policy::device::StateKind;
 use embedded_services::power::policy::{self, action};
 use embedded_services::type_c::controller::{self, Contract, Controller, PortStatus};
 use embedded_services::type_c::event::{PortEventFlags, PortEventKind};
 use embedded_services::{error, info, intrusive_list, trace, warn};
 use embedded_usb_pd::{Error, PdError, PortId as LocalPortId};
+
+mod power;
 
 /// Takes an implementation of the `Controller` trait and wraps it with logic to handle
 /// message passing and power-policy integration.
@@ -34,14 +36,6 @@ impl<'a, const N: usize, C: Controller> ControllerWrapper<'a, N, C> {
         }
     }
 
-    /// Return the power device for the given port
-    fn get_power_device<'b>(&'b self, port: LocalPortId) -> Result<&'b policy::device::Device, Error<C::BusError>> {
-        if port.0 > N as u8 {
-            return PdError::InvalidPort.into();
-        }
-        Ok(&self.power[port.0 as usize])
-    }
-
     /// Handle a plug event
     /// None of the event processing functions return errors to allow processing to continue for other ports on a controller
     async fn process_plug_event(&self, power: &policy::device::Device, status: &PortStatus) {
@@ -62,72 +56,16 @@ impl<'a, const N: usize, C: Controller> ControllerWrapper<'a, N, C> {
             if let Ok(state) = power.try_device_action::<action::Detached>().await {
                 if let Err(e) = state.attach().await {
                     error!("Error attaching power device: {:?}", e);
-                    return;
                 }
             } else {
                 // This should never happen
                 error!("Power device not in detached state");
-                return;
             }
         } else {
             info!("Plug removed");
             if let Err(e) = power.detach().await {
                 error!("Error detaching power device: {:?}", e);
-                return;
             };
-        }
-    }
-
-    /// Handle a new consumer contract
-    /// None of the event processing functions return errors to allow processing to continue for other ports on a controller
-    async fn process_new_consumer_contract(&self, power: &policy::device::Device, status: &PortStatus) {
-        info!("New consumer contract");
-
-        if let Some(contract) = status.contract {
-            if !matches!(contract, Contract::Sink(_)) {
-                error!("Not a sink contract");
-                return;
-            }
-        } else {
-            error!("No contract");
-            return;
-        }
-
-        let contract = status.contract.unwrap();
-        let current_state = power.state().await.kind();
-        // Don't update the available consumer contract if we're providing power
-        if current_state != StateKind::ConnectedProvider {
-            // Recover if we're not in the correct state
-            match power.device_action().await {
-                action::device::AnyState::Detached(state) => {
-                    if let Err(e) = state.attach().await {
-                        error!("Error attaching power device: {:?}", e);
-                        return;
-                    }
-                }
-                _ => {}
-            }
-
-            if let Ok(state) = power.try_device_action::<action::Idle>().await {
-                if let Err(e) = state
-                    .notify_consumer_power_capability(Some(policy::PowerCapability::from(contract)))
-                    .await
-                {
-                    error!("Error setting power contract: {:?}", e);
-                    return;
-                }
-            } else if let Ok(state) = power.try_device_action::<action::ConnectedConsumer>().await {
-                if let Err(e) = state
-                    .notify_consumer_power_capability(Some(policy::PowerCapability::from(contract)))
-                    .await
-                {
-                    error!("Error setting power contract: {:?}", e);
-                    return;
-                }
-            } else {
-                error!("Power device not in detached state");
-                return;
-            }
         }
     }
 
@@ -193,52 +131,8 @@ impl<'a, const N: usize, C: Controller> ControllerWrapper<'a, N, C> {
         self.pd_controller.notify_ports(port_events).await;
     }
 
-    /// Wait for a power command
-    async fn wait_power_command(&self) -> (RequestData, LocalPortId) {
-        let futures: [_; N] = from_fn(|i| self.power[i].wait_request());
-
-        let (command, local_id) = select_array(futures).await;
-        trace!("Power command: device{} {:#?}", local_id, command);
-        (command, LocalPortId(local_id as u8))
-    }
-
-    /// Process a power command
-    /// Returns no error because this is a top-level function
-    async fn process_power_command(&self, controller: &mut C, port: LocalPortId, command: RequestData) {
-        trace!("Processing power command: device{} {:#?}", port.0, command);
-        let power = match self.get_power_device(port) {
-            Ok(power) => power,
-            Err(_) => {
-                error!("Port{}: Error getting power device for port", port.0);
-                return;
-            }
-        };
-
-        match command {
-            policy::device::RequestData::ConnectConsumer(capability) => {
-                info!("Port{}: Connect consumer: {:?}", port.0, capability);
-                if let Err(_) = controller.enable_sink_path(port, true).await {
-                    error!("Error enabling sink path");
-                    power.send_response(Err(policy::Error::Failed)).await;
-                    return;
-                }
-            }
-            policy::device::RequestData::Disconnect => {
-                info!("Port{}: Disconnect", port.0);
-                if let Err(_) = controller.enable_sink_path(port, false).await {
-                    error!("Error disabling sink path");
-                    power.send_response(Err(policy::Error::Failed)).await;
-                    return;
-                }
-            }
-            _ => {}
-        }
-
-        power.send_response(Ok(policy::device::ResponseData::Complete)).await;
-    }
-
     /// Top-level processing function
-    ///
+    #[allow(clippy::await_holding_refcell_ref)]
     pub async fn process(&self) {
         let mut controller = self.controller.borrow_mut();
         match select(controller.wait_port_event(), self.wait_power_command()).await {
